Make the frame buffer size to 2:1 (#2)
The problem is the buffer scaling in the DVI output handler function.
It causes the delays and make broken output.

To fix this problem, I make buffer scaling ratio to 1:2, to remove the scaling loop in the DVI output handler.
Thus, the buffers have full width and half the height of the output size.
Memory usage with double buffering matchies the display size (640 x 480 = 307,200B).